first commit
This code is buggy. I have a hard time with reader the user(lsp) input correctly.
In fact, the offender here is the custom 'bufio.SplitFunc' I have provided. It do not what it is suppose to.
After spending a few days on this bug, my hypothesis is that when the custom split function return an 'advance' count
greater than zero, it should as well return a data. Otherwise, there is funny bug.

For instance this is okay
	return index + 4, data[:index], nil

But this is not okay :
	return index + 4, nil, nil

First the definition of the splitFunc is as follow: func (data []byte, atEOF bool) (advance int, token []byte, err error)
Then, the first example is good since it return both an 'advance' and a token (data[:index]).
On the other hand, the second is bad because it 'advance' but it doesn't return a token (nil) to the use, which is not
recommended since some bug may creep up.